Show worker progress.
This change switches back to using the "Compat" fragments, which allows
observing the LiveData<WorkInfo> reported by the WorkManager. This
WorkInfo is the basis for showing the progress UI, in the
WorkProgressFragment, and reacting to the work being finished in the
preferences.

The previous issue with the "Compat" fragments crashing at startup is
fixed by changing the theme from Theme.Leanback to
Theme.AppCompat.Leanback.